Added Api and Redis Support for CardAttributeHelper

+++ PkmnCardgameBattleCalculatorDemo/BattleCalculatorDemo.Models/CombatResult.cs
@@ -2,7 +2,7 @@ namespace BattleCalculatorDemo.Models
 {
     public class CombatResult
     {
-        public short DamageDealt { get; set; }
+        public int DamageDealt { get; set; }
         public bool WasCritical { get; set; }
         public bool DidDefenderDie { get; set; } = false;
     }
